+++ src/Products/PloneMeeting/adapters.py
@@ -27,9 +27,12 @@ from imio.helpers.cache import get_plone_groups_for_user
 from imio.helpers.catalog import merge_queries
 from imio.helpers.content import get_vocab
 from imio.helpers.content import get_vocab_values
+from imio.helpers.content import richtextval
 from imio.helpers.xhtml import xhtmlContentIsEmpty
 from imio.history.adapters import BaseImioHistoryAdapter
 from imio.history.adapters import ImioWfHistoryAdapter
+from imio.history.interfaces import IImioHistory
+from imio.history.utils import getLastAction
 from imio.prettylink.adapters import PrettyLinkAdapter
 from persistent.list import PersistentList
 from plone import api
@@ -61,12 +64,14 @@ from Products.PloneMeeting.utils import compute_item_roles_to_assign_to_suffixes
 from Products.PloneMeeting.utils import displaying_available_items
 from Products.PloneMeeting.utils import findNewValue
 from Products.PloneMeeting.utils import get_context_with_request
+from Products.PloneMeeting.utils import get_dx_attrs
 from Products.PloneMeeting.utils import get_referer_obj
 from Products.PloneMeeting.utils import getCurrentMeetingObject
 from Products.PloneMeeting.utils import getHistoryTexts
 from Products.PloneMeeting.utils import is_transition_before_date
 from z3c.form.term import MissingChoiceTermsVocabulary
 from zope.annotation import IAnnotations
+from zope.component import getAdapter
 from zope.i18n import translate
 from zope.schema.vocabulary import SimpleVocabulary
 
@@ -119,7 +124,7 @@ class AdviceContentDeletableAdapter(APContentDeletableAdapter):
     """
       Manage the mayDelete for meetingadvice.
       Must have 'Delete objects' on the item.
-      If some versions are saved (advice was asked_again at least once), advice
+      If advice was historized (advice was asked_again at least once), advice
       is not deletable.
     """
     def __init__(self, context):
@@ -131,10 +136,9 @@ class AdviceContentDeletableAdapter(APContentDeletableAdapter):
         mayDelete = super(AdviceContentDeletableAdapter, self).mayDelete()
         if mayDelete:
             tool = api.portal.get_tool('portal_plonemeeting')
-            pr = api.portal.get_tool('portal_repository')
             if not tool.isManager(realManagers=True) and \
-               pr.getHistoryMetadata(self.context):
-                return False
+               getLastAction(getAdapter(self.context, IImioHistory, 'advice_given')):
+                mayDelete = False
         return mayDelete
 
 
@@ -790,6 +794,25 @@ class PMCompletenessChangesHistoryAdapter(BaseImioHistoryAdapter):
 
     history_type = 'completeness_changes'
     history_attr_name = 'completeness_changes_history'
+
+
+class PMAdviceGivenHistoryAdapter(BaseImioHistoryAdapter):
+    """ """
+
+    history_type = 'advice_given'
+    history_attr_name = 'advice_given_history'
+
+    def revert_to_last_event(self):
+        """Revert advice values to last historized event."""
+        last_action = getLastAction(self)
+        rich_text_field_names = get_dx_attrs(
+            self.context.portal_type, richtext_only=True, as_display_list=False)
+        for field_data in last_action['advice_data']:
+            # handle rich text to store a RichTextValue
+            field_value = field_data['field_value']
+            if field_data['field_name'] in rich_text_field_names:
+                field_value = richtextval(field_value)
+            setattr(self.context, field_data['field_name'], field_value)
 
 
 class Criteria(eeaCriteria):