+++ src/Products/PloneMeeting/utils.py
@@ -6,6 +6,7 @@ from Acquisition import aq_base
 from appy.pod.xhtml2odt import XhtmlPreprocessor
 from appy.shared.diff import HtmlDiff
 from bs4 import BeautifulSoup
+from collections import OrderedDict
 from collective.behavior.talcondition.utils import _evaluateExpression
 from collective.contact.core.utils import get_gender_and_number
 from collective.contact.core.utils import get_position_type_name
@@ -39,6 +40,9 @@ from imio.helpers.xhtml import replace_content
 from imio.helpers.xhtml import separate_images
 from imio.helpers.xhtml import storeImagesLocally
 from imio.helpers.xhtml import xhtmlContentIsEmpty
+from imio.history.interfaces import IImioHistory
+from imio.history.utils import add_event_to_history
+from imio.history.utils import getLastAction
 from imio.history.utils import getLastWFAction
 from plone import api
 from plone.app.textfield import RichText
@@ -941,7 +945,7 @@ def get_dx_attrs(portal_type,
     # FIELD_INFOS
     portal_types = api.portal.get_tool('portal_types')
     fti = portal_types[portal_type]
-    field_infos = resolveDottedName(fti.klass).FIELD_INFOS
+    field_infos = getattr(resolveDottedName(fti.klass), "FIELD_INFOS", {})
     for field_name, field in schema_fields:
         if optional_only and \
            (field_name not in field_infos or not field_infos[field_name]['optional']):
@@ -1010,6 +1014,32 @@ def get_dx_widget(obj, field_name, mode=DISPLAY_MODE):
     # this will set widget.__name__
     locate(widget, None, field_name)
     return widget
+
+
+def get_dx_data(obj):
+    """ """
+    data = []
+    # use print_value available on the documentgenerator helper view
+    view = obj.unrestrictedTraverse('@@document-generation')
+    helper = view.get_generation_context_helper()
+    for attr_name in get_dx_attrs(obj.portal_type):
+        field_content = None
+        try:
+            field_content = helper.print_value(attr_name, raw_xhtml=True)
+        except Exception:
+            logger.warning(
+                "In \"utils.get_dx_data\", could not print_value for attr_name "
+                "\"%s\" with value \"%s\" for element at \"%s\"" %
+                (attr_name, getattr(obj, attr_name), "/".join(obj.getPhysicalPath())))
+        # do not store a RichTextValue, store the rendered value
+        field_value = getattr(obj, attr_name)
+        if isinstance(field_value, RichTextValue):
+            field_value = field_content
+        data.append(
+            {'field_name': attr_name,
+             'field_value': field_value,
+             'field_content': field_content})
+    return data
 
 
 def set_dx_value(obj, field_name, value, raise_unauthorized=True):
@@ -1561,18 +1591,13 @@ def get_context_with_request(context):
 
 
 def isModifiedSinceLastVersion(obj):
-    """Check if given p_obj was modified since last version (versioning)."""
-    pr = api.portal.get_tool('portal_repository')
-    history_metadata = pr.getHistoryMetadata(obj)
+    """Check if given p_obj was modified since last version (history)."""
+    adapter = getAdapter(obj, IImioHistory, 'advice_given')
+    last_event = getLastAction(adapter)
     modified = True
-    if history_metadata and history_metadata._available:
-        # date it was versionned
-        timestamp = history_metadata._full[history_metadata.nextVersionId - 1]['metadata']['sys_metadata']['timestamp']
-        # we do not use _retrieve because it does a transaction savepoint and it
-        # breaks collective.zamqp...  So we use timestamp
-        # advice.modified will be older than timestamp as it is managed in see content.advice.versionate_if_relevant
+    if last_event:
         # keep >= for backward compatibility as before, modified was set to timestamp, now it is older...
-        if DateTime(timestamp) >= obj.modified():
+        if last_event['time'] >= obj.modified():
             modified = False
     return modified
 
@@ -1591,6 +1616,21 @@ def version_object(obj, keep_modified=True, only_once=False, comment=''):
     pr.save(obj=obj, comment=comment)
     # set back modified on obj so version timestamp is > obj modified
     obj.setModificationDate(obj_modified)
+
+
+def historize_object_data(obj, comment):
+    """Historize p_obj data and parent's data.
+       An optionnal p_comment may be defined and will be usedappear in the versions history."""
+    # compute advice data and item data
+    item_data = main_item_data(obj.aq_parent)
+    advice_data = get_dx_data(obj)
+    add_event_to_history(
+        obj,
+        'advice_given_history',
+        action='advice_given_or_modified',
+        comments=comment,
+        extra_infos={'item_data': item_data,
+                     'advice_data': advice_data})
 
 
 # taken from http://mscerts.programming4.us/fr/639402.aspx